#23 Refactor(Utils): 프로젝트 전체에서 사용할 util 파일 생성
core 디렉토리의 utils.py 파일을 생성하였음.

프로젝트 전체에서 사용할 함수들을 utils.py 파일에 작성할 것.

JWT 인증 함수인 authenticate_token 함수를 utils.py 파일로 이동하였음.
