+++ backend/account/views.py
@@ -3,19 +3,17 @@ from django.conf import settings
 from django.http import JsonResponse, HttpResponseRedirect
 from django.shortcuts import redirect
 from django.contrib.auth import get_user_model
-from rest_framework_simplejwt.tokens import AccessToken, RefreshToken
+from rest_framework_simplejwt.tokens import RefreshToken
 from rest_framework.decorators import api_view
 from django.contrib.auth.hashers import make_password, check_password
 from django.core.mail import send_mail
-from rest_framework_simplejwt.exceptions import TokenError
 from django.views.generic import View
 from django.middleware.csrf import get_token
 from django.views.decorators.csrf import ensure_csrf_cookie
-from django.views.decorators.csrf import csrf_protect
-from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_exempt
 from django.core.files.storage import default_storage
 from django.core.files.base import ContentFile
+from core.utils import authenticate_token
 
 User = get_user_model()
 
@@ -492,72 +490,6 @@ def local_auth_sign_up(request):
         )
     except Exception as e:
         return JsonResponse({"message": f"{str(e)}"}, status=500)
-
-
-def authenticate_token(request):
-    """
-    @brief JWT 검증 함수
-
-    @param request 그대로 넘겨줌.
-
-    @return
-        - 성공 : (유저 객체, None) 형식으로 반환
-        - 실패 (access_token 만료) : (None, 토큰 갱신 메시지) 형식으로 반환
-        - 실패 (refresh_token 만료 or 유효하지 않은 JWT) : (None, 실패 메시지) 형식으로 반환
-
-    @details
-    request의 쿠키에서 access_token과 refresh_token을 가져온다.
-    access_token을 디코딩하여 user_id를 추출한다.
-    access_token의 payload에서 user_id(pk)를 가져온다.
-        - access_token이 없거나 만료된 경우, refresh_token을 사용해 새로운 access_token을 생성하고 (None, 토큰 갱신 메시지) 형식으로 반환한다.
-        - refresh_token이 만료되었거나 access_token이 유효하지 않으면 (None, 실패 메시지) 형식으로 반환한다.
-    추출된 user_id를 기반으로 User 테이블에서 사용자 객체를 찾아서 (User, None) 형식으로 반환한다.
-    """
-    access_token = request.COOKIES.get("access_token")
-    refresh_token = request.COOKIES.get("refresh_token")
-
-    if not access_token and not refresh_token:
-        return None, JsonResponse({"message": "JWTs are missing."}, status=401)
-
-    user_id = None
-
-    if access_token:
-        try:
-            access_payload = AccessToken(access_token)
-            user_id = access_payload.get("user_id")
-        except TokenError:
-            pass
-
-    if not user_id:
-        try:
-            refresh = RefreshToken(refresh_token)
-            new_access = refresh.access_token
-
-            response = JsonResponse({"message": "Token refreshed"}, status=452)
-            response.set_cookie(
-                key="access_token",
-                value=str(new_access),
-                httponly=True,
-                secure=True,
-                samesite="Lax",
-                max_age=settings.SIMPLE_JWT["ACCESS_TOKEN_LIFETIME"].total_seconds(),
-                path="/",
-            )
-            return None, response
-        except TokenError:
-            return None, JsonResponse(
-                {"message": "Invalid or expired refresh token"}, status=401
-            )
-
-    if not user_id:
-        return None, JsonResponse({"message": "Invalid token payload"}, status=401)
-
-    try:
-        user = User.objects.get(id=user_id)
-    except User.DoesNotExist:
-        return None, JsonResponse({"message": "User not found"}, status=401)
-
-    return user, None
 
 
 @api_view(["POST"])
